backend/requirements.txt
# Core
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1

# Database
SQLAlchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0

# Auth
Werkzeug==3.1.3
google-auth==2.37.0
google-auth-oauthlib==1.2.1
PyJWT==2.10.1
cryptography==44.0.0

# OCR
google-cloud-vision==3.9.0

# HTTP / Utils
requests==2.32.3
python-dotenv==1.0.1
python-dateutil==2.9.0
Pillow==11.1.0

# Validation
marshmallow==3.23.1

# Firebase / Push
firebase-admin==6.7.0

# Storage
boto3==1.36.7

# Testing
pytest==8.3.4
pytest-flask==1.3.0
pytest-cov==6.0.0
factory-boy==3.3.1

# Scheduler
flask-apscheduler==1.13.1
apscheduler==3.11.2

# Email
resend==2.27.0

# Production
gunicorn==23.0.0
